backend/requirements.txt
fastapi==0.108.0
uvicorn[standard]==0.25.0
sqlalchemy==2.0.25
alembic==1.13.1
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
yfinance==0.2.33
alpha-vantage==2.3.1
openai==1.6.1
apscheduler==3.10.4
python-dotenv==1.0.0
httpx==0.26.0
pandas>=2.1.0
numpy>=1.24.0
pytz==2023.3
pytest==7.4.4
pytest-asyncio==0.21.1
setuptools>=65.0.0
google-auth==2.25.2
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1